translation/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
--find-links https://download.pytorch.org/whl/torch_stable.html

certifi==2021.5.30
    # via requests
charset-normalizer==2.0.4
    # via requests
click==8.0.1
    # via
    #   sacremoses
    #   typer
filelock==3.0.12
    # via
    #   huggingface-hub
    #   transformers
huggingface-hub==0.0.12
    # via transformers
idna==3.2
    # via requests
joblib==1.0.1
    # via sacremoses
kafka-python==2.0.2
    # via -r requirements.in
numpy==1.21.1
    # via transformers
packaging==21.0
    # via
    #   huggingface-hub
    #   transformers
protobuf==3.17.3
    # via transformers
pydantic==1.8.2
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pyyaml==5.4.1
    # via transformers
redis==3.5.3
    # via -r requirements.in
regex==2021.8.3
    # via
    #   sacremoses
    #   transformers
requests==2.26.0
    # via
    #   huggingface-hub
    #   transformers
sacremoses==0.0.45
    # via transformers
sentencepiece==0.1.91
    # via transformers
six==1.16.0
    # via
    #   protobuf
    #   sacremoses
tokenizers==0.10.3
    # via transformers
torch==1.9.0+cu111
    # via -r requirements.in
tqdm==4.62.0
    # via
    #   huggingface-hub
    #   sacremoses
    #   transformers
transformers[sentencepiece]==4.9.1
    # via -r requirements.in
typer==0.4.0
    # via -r requirements.in
typing-extensions==3.10.0.0
    # via
    #   huggingface-hub
    #   pydantic
    #   torch
urllib3==1.26.6
    # via requests
